Modified the PackageEntity to support the newly added licensing scheme. The license information is stored in CKAN and populated when the package is fetched
